SocRATES/requirements.txt
ament_copyright==0.12.11
ament_flake8==0.12.11
ament_index_python==1.4.0
ament_pep257==0.12.11
click==8.1.3
cv_bridge==3.2.1
geometry_msgs==4.2.4
google_auth==2.29.0
google_cloud_aiplatform==1.48.0
ipython==8.12.3
ipywidgets==8.0.4
launch_ros==0.19.7
networkx==3.2.1
numpy==2.0.1
omegaconf==2.3.0
openai==1.37.1
opencv_python==4.8.0.76
opencv_python_headless==4.8.0.76
Pillow==10.4.0
pytest==7.4.0
PyYAML==6.0.1
PyYAML==6.0.1
rclpy==3.3.13
scripts==3.7.0
sensor_msgs==4.2.4
setuptools==68.2.2
setuptools==68.0.0
std_srvs==4.2.4
tenacity==8.2.2
termcolor==2.4.0
tiktoken==0.5.2
tqdm==4.66.1
